Adjust files to different components
