+++ src/components/SlideList.jsx
@@ -1,29 +1,26 @@
 import PropTypes from 'prop-types';
 
-const SlideView = ({selectedSlide, slides}) => {
-
-    const displaySlide = (slide) => (
-        slide.map((element, idx) => (
-          <div key={idx}>
-            <h3> {element.type}:</h3>
-            <p>{element.value}</p>
-          </div>
-        ))
-      );
-   return (
-   <div className="col-span-2 border border-gray-300 p-4 rounded-lg"> {/* middle column */}
-    <div className="flex justify-center w-full">
-      <div className="w-3/4">
-        {selectedSlide !== null && slides[selectedSlide] ? displaySlide(slides[selectedSlide]) : null }
-      </div>
+const SlideList = ({ setSelectedSlide, slides }) => {
+  return (
+    <div className="col-span-1">
+      {slides.map((slide, slideIndex) => (
+        slide.length > 0 && slide[0].type === 'title' ? (
+          <button
+            className="block text-left"
+            onClick={() => setSelectedSlide(slideIndex)}
+            key={slideIndex} // Use slideIndex as the key
+          >
+            {slide[0].value}
+          </button>
+        ) : null
+      ))}
     </div>
-  </div>)
-}
+  );
+};
 
-SlideView.propTypes = {
-    slides: PropTypes.array.isRequired,
-    selectedSlide: PropTypes.array.isRequired,
-  };
+SlideList.propTypes = {
+  slides: PropTypes.array.isRequired,
+  setSelectedSlide: PropTypes.array,
+};
 
-export default SlideView;
-
+export default SlideList;
