+++ src/components/SlideView.jsx
@@ -1,17 +1,29 @@
+import PropTypes from 'prop-types';
 
-const SlideView = ({setSelectedSlide, slides}) => {
-    <div className="col-span-1"> 
-    {
-      slides.map((slide, slideIndex) => (
-        slide.length > 0 && slide[0].type == 'title' ? (
-          <button className="block text-left" onClick={() => setSelectedSlide(slideIndex)} key={slide[0].title}> 
-            {slide[0].value}
-          </button>
-        ) : null
-      ))
-    }
-  </div>
+const SlideView = ({selectedSlide, slides}) => {
+
+    const displaySlide = (slide) => (
+        slide.map((element, idx) => (
+          <div key={idx}>
+            <h3> {element.type}:</h3>
+            <p>{element.value}</p>
+          </div>
+        ))
+      );
+   return (
+   <div className="col-span-2 border border-gray-300 p-4 rounded-lg"> {/* middle column */}
+    <div className="flex justify-center w-full">
+      <div className="w-3/4">
+        {selectedSlide !== null && slides[selectedSlide] ? displaySlide(slides[selectedSlide]) : null }
+      </div>
+    </div>
+  </div>)
 }
+
+SlideView.propTypes = {
+    slides: PropTypes.array.isRequired,
+    selectedSlide: PropTypes.array,
+  };
 
 export default SlideView;
